# public/openapi/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DebtResolve API",
        "contact": {
            "email": "jadiael@hotmail.com.br"
        },
        "license": {
            "name": "CC-BY-NC",
            "url": "https://creativecommons.org/licenses/by-nc/4.0/"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/users/installments": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "List Installments for User",
                "description": "Retrieve a list of installments for the authenticated user",
                "operationId": "fbc8633c1e8dac8499c12bf0b3827f00",
                "responses": {
                    "200": {
                        "description": "List of installments for the user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "List of installments for this user"
                                        },
                                        "data": {
                                            "properties": {
                                                "installments": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Installment"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        },
                                        "errors": {
                                            "type": "null"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 401
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/users/charges": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "List Charges associated with User",
                "description": "Retrieve a list of charges associated with the authenticated user",
                "operationId": "ea2d0ea0661892dcd0d81d902ebab36c",
                "responses": {
                    "200": {
                        "description": "List of charges for the user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Charges associated with this user"
                                        },
                                        "data": {
                                            "properties": {
                                                "debtors": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/User"
                                                    }
                                                },
                                                "collectors": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/User"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        },
                                        "errors": {
                                            "type": "null"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 401
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "List Users",
                "description": "Retrieve a list of all users",
                "operationId": "7270627e71b74db2cd48a7618947b439",
                "responses": {
                    "200": {
                        "description": "Users successfully listed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Users successfully listed"
                                        },
                                        "data": {
                                            "properties": {
                                                "users": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/User"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/users/{user}/user": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Show User",
                "description": "Show details of a specific user",
                "operationId": "33571a4e35cbeff4932ce480a5c16d93",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User successfully listed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User successfully listed"
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "$ref": "#/components/schemas/User"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User not found"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Charge": {
                "properties": {
                    "id": {
                        "description": "Unique identifier for the charge",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Name of the charge",
                        "type": "string",
                        "example": "Monthly rent"
                    },
                    "description": {
                        "description": "Description of the charge",
                        "type": "string",
                        "example": "Rent payment for the month"
                    },
                    "amount": {
                        "description": "Total amount of the charge",
                        "type": "number",
                        "example": 500
                    },
                    "installments_number": {
                        "description": "Number of installments for the charge",
                        "type": "integer",
                        "example": 12
                    },
                    "due_day": {
                        "description": "Due day for payment",
                        "type": "integer",
                        "example": 15
                    },
                    "collector_id": {
                        "description": "ID of the user who is the collector for this charge",
                        "type": "integer",
                        "example": 1
                    },
                    "debtor_id": {
                        "description": "ID of the user who is the debtor for this charge",
                        "type": "integer",
                        "example": 2
                    },
                    "collector": {
                        "$ref": "#/components/schemas/User"
                    },
                    "debtor": {
                        "$ref": "#/components/schemas/User"
                    },
                    "installments": {
                        "description": "List of installments associated with this charge",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Installment"
                        }
                    }
                },
                "type": "object"
            },
...
            "Installment": {
                "properties": {
                    "id": {
                        "description": "Unique identifier for the installment",
                        "type": "integer",
                        "example": 1
                    },
                    "value": {
                        "description": "Value of the installment",
                        "type": "number",
                        "example": 100
                    },
                    "installment_number": {
                        "description": "Number of the installment",
                        "type": "integer",
                        "example": 1
                    },
                    "due_date": {
                        "description": "Due date for payment",
                        "type": "string",
                        "format": "date",
                        "example": "2023-11-30"
                    },
                    "paid": {
                        "description": "Indicates if the installment is paid or not",
                        "type": "boolean",
                        "example": true
                    },
                    "user_id": {
                        "description": "ID of the user associated with the installment",
                        "type": "integer",
                        "example": 1
                    },
                    "awaiting_approval": {
                        "description": "Indicates if the installment is awaiting approval",
                        "type": "boolean",
                        "example": false
                    },
                    "payment_proof": {
                        "description": "URL or path to the payment proof file",
                        "type": "string",
                        "example": "https://example.com/payment-proof.jpg"
                    },
                    "charge": {
                        "$ref": "#/components/schemas/Charge"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "integer"
                    },
                    "name": {
                        "description": "User's name",
                        "type": "string"
                    },
                    "email": {
                        "description": "User's email address",
                        "type": "string",
                        "format": "email"
                    },
                    "email_verified_at": {
                        "description": "Date/time when email was verified",
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "description": "Date/time of user creation",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date/time of user update",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "Bearer"
            }
        }
    },
...
}